perf: remove unnecessary variable in methodList
docs: improve wording in natspec
chore: various improvements in comments
test: change the header for the target functions in Base_Test
test: cast to uint256 in assertEq function for the Status

+++ test/utils/Assertions.sol
@@ -21,11 +21,11 @@ abstract contract Assertions is PRBTest {
 
     /// @dev Compares two `Lockup.Status` enum values.
     function assertEq(Lockup.Status a, Lockup.Status b) internal {
-        assertEq(uint8(a), uint8(b), "status");
+        assertEq(uint256(a), uint256(b), "status");
     }
 
     /// @dev Compares two `Lockup.Status` enum values.
     function assertEq(Lockup.Status a, Lockup.Status b, string memory err) internal {
-        assertEq(uint8(a), uint8(b), err);
+        assertEq(uint256(a), uint256(b), err);
     }
 }
